Geo Grids

There are now 216 grids available in this package and more online. To view a full list of available grids, see here. To create and submit your own grid, see here. To see several examples of grids being used to visualize data, see facet_geo.

us_state_grid1: Grid layout for US states (including DC).

us_state_grid2: Grid layout for US states (including DC).

eu_grid1: Grid layout for the 28 EU Countries.

aus_grid1: Grid layout for the Australian States and Territories.. Thanks to jonocarroll.

sa_prov_grid1: Grid layout for the provinces of South Africa. Thanks to jonmcalder.

gb_london_boroughs_grid: Grid layout for the boroughs of London. Note that the column code_ons contains the codes used by UK Office for National Statistics.. Thanks to eldenvo.

nhs_scot_grid: Grid layout for a grid of NHS Scotland Health Boards. Note that the column code contains the codes used by UK Office for National Statistics.. Thanks to jsphdms.

india_grid1: Grid layout for India states (not including union territories).. Thanks to meysubb.

india_grid2: Grid layout for India states (not including union territories)..

argentina_grid1: Grid for the 23 provinces of Argentina. It includes the Malvinas/Falkland Islands and the Antarctic Territories (these are disputed, but they are included since many researchers might use data from these locations).. Thanks to eliocamp.

br_states_grid1: Grid for the 27 states of Brazil.. Thanks to italocegatta.

sea_grid1: Grid for South East Asian countries.. Thanks to jasonjb82.

mys_grid1: Grid for Malaysian states and territories.. Thanks to jasonjb82.

fr_regions_grid1: Land and overseas regions of France. Codes are INSEE codes.. Thanks to mtmx.

de_states_grid1: Grid for the German states ('Länder'). Thanks to DominikVogel.

us_or_counties_grid1: Grid for Oregon counties.. Thanks to aosmith16.

us_wa_counties_grid1: Grid for Washington counties..

us_in_counties_grid1: Grid for Indiana counties.. Thanks to nateapathy.

us_in_central_counties_grid1: Grid for central Indiana counties.. Thanks to nateapathy.

se_counties_grid1: Grid for counties of Sweden.. Thanks to duleise.

sf_bay_area_counties_grid1: Grid of the 9 San Francisco Bay Area counties.. Thanks to Eunoia.

ua_region_grid1: Grid of administrative divisions of Ukraine (24 oblasts, one autonomous region, and two cities).. Thanks to woldemarg.

mx_state_grid1: Grid layout for the states of Mexico.. Thanks to ikashnitsky.

mx_state_grid2: Grid layout for the states of Mexico.. Thanks to diegovalle.

scotland_local_authority_grid1: Grid layout for the local authorities of Scotland.. Thanks to davidhen.

us_state_without_DC_grid1: Grid layout for US states (excluding DC). Thanks to ejr248.

italy_grid1: Grid layout for regions of Italy (in collaboration with Stella Cangelosi and Luciana Dalla Valle).. Thanks to JulianStander.

italy_grid2: Grid layout for regions of Italy (in collaboration with Stella Cangelosi and Luciana Dalla Valle).. Thanks to JulianStander.

be_province_grid1: Grid layout for provinces of Belgium plus Brussels, including names in three languages (French, Dutch, English) and Belgium internal codes (NIS).. Thanks to ericlecoutre.

us_state_grid3: Grid layout for US states (including DC).. Thanks to kanishkamisra.

jp_prefs_grid1: Grid layout for the prefectures of Japan.. Thanks to uribo.

ng_state_grid1: Grid layout for the 37 Federal States of Nigeria.. Thanks to aledemogr.

bd_upazila_grid1: Grid layout for Bangladesh 64 Upazilas.. Thanks to aledemogr.

spain_prov_grid1: Grid layout for Provinces of Spain.. Thanks to kintero.

ch_cantons_grid1: Grid layout for Cantons of Switzerland.. Thanks to tinu-schneider.

ch_cantons_grid2: Grid layout for Cantons of Switzerland.. Thanks to rastrau.

china_prov_grid1: Grid layout for Provinces of China.. Thanks to weiyunna.

world_86countries_grid: Grid layout for 86 countries in the world.. Thanks to akangsha.

se_counties_grid2: Grid for counties of Sweden.. Thanks to richardohrvall.

uk_regions1: Grid for regions of the UK (aka EU standard NUTS 1 areas).. Thanks to paulb20.

us_state_contiguous_grid1: Grid layout for the contiguous US states (including DC).. Thanks to andrewsr.

sk_province_grid1: Grid layout for South Korean sis and dos (metropolitan/special/autonomous cities and provinces).. Thanks to heon131.

ch_aargau_districts_grid1: Grid layout for Districts of the Canton of Aargau, Switzerland.. Thanks to zumbov2.

jo_gov_grid1: Grid layout for Governorates of Jordan.. Thanks to aledemogr.

es_autonomous_communities_grid1: Grid layout for Spanish 'Comunidades Autónomas'.. Thanks to JoseAntonioOrtega.

spain_prov_grid2: Grid layout for Provinces of Spain.. Thanks to JoseAntonioOrtega.

world_countries_grid1: Grid layout for countries of the world, with a few exclusions. See .. Thanks to JoseAntonioOrtega.

br_states_grid2: Grid for the 27 states of Brazil.. Thanks to hafen.

china_city_grid1: Grid layout of cities in China.. Thanks to CharleneDeng1.

kr_seoul_district_grid1: Grid layout of Seoul's 25 districts.. Thanks to yonghah.

nz_regions_grid1: Grid layout for regions of New Zealand.. Thanks to pierreroudier.

sl_regions_grid1: Grid layout of Slovenian regions.. Thanks to SR1986.

us_census_div_grid1: Grid layout of US Census divisions.. Thanks to mkiang.

ar_tucuman_province_grid1: Grid layout for Argentina Tucumán Province political divisions (departments). Thanks to TuQmano.

us_nh_counties_grid1: Grid layout for the 10 counties in New Hampshire.. Thanks to ghost.

china_prov_grid2: Grid layout for Provinces of China.. Thanks to jw2531.

pl_voivodeships_grid1: Grid layout for Polish voivodeships (provinces). Thanks to erzk.

us_ia_counties_grid1: Grid layout for counties in Iowa. Thanks to jrennyb.

us_id_counties_grid1: Grid layout for counties in Idaho. Thanks to hathawayj.

ar_cordoba_dep_grid1: Grid layout for departments of Cordoba province in Argentina.. Thanks to TuQmano.

us_fl_counties_grid1: Grid for Florida counties.. Thanks to ejr248.

ar_buenosaires_communes_grid1: Grid for communes of Buenos Aires, Argentina.. Thanks to TuQmano.

nz_regions_grid2: Grid layout for regions of New Zealand.. Thanks to pierreroudier.

oecd_grid1: Grid layout for OECD member countries.. Thanks to arcruz0.

ec_prov_grid1: Grid layout for provinces of Ecuador. Thanks to Ricardo95RM.

nl_prov_grid1: Grid layout for provinces of Netherlands. Thanks to ruditurksema.

ca_prov_grid1: Grid layout for provinces of Canada. Thanks to michael-chong.

us_nc_counties_grid1: Grid layout for Counties of North Carolina, United States. Thanks to mtdukes.

mx_ciudad_prov_grid1: Grid layout for Districts of Mexico City, Mexico. Thanks to Ivangea.

bg_prov_grid1: Grid layout for provinces of Bulgaria. Thanks to savinastoitsova.

us_hhs_regions_grid1: This grid approximates the U.S. Health and Human Services Region map. Thanks to akitepowell.

tw_counties_grid1: Grid layout for Counties of Taiwan. Thanks to csh484912274.

tw_counties_grid2: Grid layout for Counties of Taiwan including Lienchiang County. Thanks to csh484912274.

af_prov_grid1: Grid layout for Provinces of Afghanistan. Thanks to jrennyb.

us_mi_counties_grid1: Grid layout for Counties of Michigan, United States. Thanks to jrennyb.

pe_prov_grid1: Grid layout for Provinces of Peru. Thanks to jmcastagnetto.

sa_prov_grid2: Grid layout for Provinces of South Africa. Thanks to kamermanpr.

mx_state_grid3: Grid layout for States of Mexico. Thanks to ikashnitsky.

cn_bj_districts_grid1: Grids for Administrative Districts of Beijing, China. Thanks to shiedelweiss.

us_va_counties_grid1: Grids for Counties of Virginia, United States. Thanks to joshyazman.

us_mo_counties_grid1: Grids for Counties of Missouri, United States. Thanks to Yanqi-Xu.

cl_santiago_prov_grid1: Communes of Santiago Province, Chile. Thanks to robsalasco.

us_tx_capcog_counties_grid1: This is a grid of a 10 county planning region around Austin, Texas, United States. Thanks to mth444.

sg_planning_area_grid1: Grids for Planning Areas of Singapore. Thanks to Elenafuyi.

in_state_ut_grid1: Grid of Indian States and Union Territories. Thanks to seanangio.

cn_fujian_prov_grid1: Grid of counties of Fujian Province, China. Thanks to nannanchen333.

ca_quebec_electoral_districts_grid1: Grid of Electoral Districts of Québec, Canada. Thanks to jhroy.

nl_prov_grid2: Grid with the provinces of The Netherlands with codes that are used by the statistical institute of NL. Thanks to edwindj.

cn_bj_districts_grid2: Grid with districts of Beijing, China. Thanks to zouhx11.

ar_santiago_del_estero_prov_grid1: Grid with districts of Santiago del Estero Province, Argentina. Thanks to TuQmano.

ar_formosa_prov_grid1: Grid with districts of Formosa Province, Argentina. Thanks to TuQmano.

ar_chaco_prov_grid1: Grid with districts of Chaco Province, Argentina. Thanks to TuQmano.

ar_catamarca_prov_grid1: Grid with districts of Catamarca Province, Argentina. Thanks to TuQmano.

ar_jujuy_prov_grid1: Grid with districts of Jujuy Province, Argentina. Thanks to TuQmano.

ar_neuquen_prov_grid1: Grid with districts of Neuquen Province, Argentina. Thanks to TuQmano.

ar_san_luis_prov_grid1: Grid with districts of San Luis Province, Argentina. Thanks to TuQmano.

ar_san_juan_prov_grid1: Grid with districts of San Juan Province, Argentina. Thanks to TuQmano.

ar_santa_fe_prov_grid1: Grid with districts of Santa Fe Province, Argentina. Thanks to TuQmano.

ar_la_rioja_prov_grid1: Grid with districts of La Rioja Province, Argentina. Thanks to TuQmano.

ar_mendoza_prov_grid1: Grid with districts of Mendoza Province, Argentina. Thanks to TuQmano.

ar_salta_prov_grid1: Grid with districts of Salta Province, Argentina. Thanks to TuQmano.

ar_rio_negro_prov_grid1: Grid with districts of Rio Negro Province, Argentina. Thanks to TuQmano.

uy_departamentos_grid1: Grid with Departamentos of Uruguay. Thanks to TuQmano.

ar_buenos_aires_prov_electoral_dist_grid1: Grid with Electoral Districts of Buenos Aires Province, Argentina. Thanks to TuQmano.

europe_countries_grid1: Grid layout for all European countries except Vatican City, Monaco, San Marino and Liechtenstein. Thanks to THargreaves.

argentina_grid2: Grid layout for Argentina without Islas Malvinas and Antártida Argentina. Thanks to TuQmano.

us_state_without_DC_grid2: Grid layout for United States with AK and HI flush with CA. Thanks to christophercannon.

jp_prefs_grid2: Grid layout for Prefectures of Japan. Thanks to Ryo-N7.

na_regions_grid1: Regions of Namibia. Thanks to stedy.

mm_state_grid1: States of Myanmar. Thanks to htinkyawaye.

us_state_with_DC_PR_grid1: United States of America including Washington, D.C. and Puerto Rico. Thanks to krothkin.

fr_departements_grid1: Grid for France's departements, the second levels of administrative boundaries after the regions. Thanks to tvroylandt.

ar_salta_prov_grid2: Grids for Salta Province Argentina. Thanks to tartagalensis.

ie_counties_grid1: Ireland counties. Code is the car number plate abbreviation for Republic counties, similar for the six counties of Northern Ireland. Tipperary is split North / South for historical reasons. Thanks to eugene100hickey.

sg_regions_grid1: Urban planning regions of Singapore. Thanks to erhuttk.

us_ny_counties_grid1: Counties of New York State, United States. Thanks to jjdfsny.

ru_federal_subjects_grid1: Federal Subjects of Russia. Thanks to ParanoidAndroid18.

us_ca_counties_grid1: Counties of the State of California, United States. Thanks to MartinLe5.

lk_districts_grid1: Second level administrative divisions of Sri Lanka. Thanks to thiyangt.

us_state_without_DC_grid3: United States grid without Washington, D.C. Thanks to ghost.

co_cali_subdivisions_grid1: Corregimientos of Cali, Columbia. Thanks to Carolina101.

us_in_northern_counties_grid1: Northern Counties of Indiana, United States. Thanks to robertoge.

italy_grid3: Autonomous Provinces of Italy. Thanks to danilolofaro.

us_state_with_DC_PR_grid2: Grid of 50 states, DC, and Puerto Rico. Thanks to krmaas.

us_state_grid7: United States grid with Washington, D.C. Thanks to yichenqin.

sg_planning_area_grid2: Singapore Planning Areas. Thanks to ZhimaoElliott.

ch_cantons_fl_grid1: Grid layout for Cantons of Switzerland and the neighbouring Prinicipality of Liechtenstein. Thanks to rastrau.

europe_countries_grid2: Grid layout for European countries (minus micro nations). Thanks to rastrau.

us_states_territories_grid1: Grid layout for U.S. states and territories. Thanks to rastrau.

us_tn_counties_grid1: Grid layout for counties of Tennesee, United States. Thanks to binkleym.

us_il_chicago_community_areas_grid1: Grid layout for the Community Areas of Chicago. Thanks to leungkp.

us_state_with_DC_PR_grid3: United States grid with Washington, D.C. and Puerto Rico. Thanks to klittle314.

in_state_ut_grid2: Grid of Indian States and Union Territories. Thanks to dnyansagar.

at_states_grid1: Grid layout for States of Austria. Thanks to wkapga.

us_pa_counties_grid1: Grid layout of Counties of Pennsylvania, United States. Thanks to urbanSpatial.

us_oh_counties_grid1: Grid layout of Counties of Ohio, United States. Thanks to taylorokonek.

fr_departements_grid2: Grid layout of Departements of France. Thanks to jerbou.

us_wi_counties_grid1: Grid layout for counties of Wisconsin, United States. Thanks to aravamu2.

africa_countries_grid1: Grid for all countries in Africa. Namibia added as 'NAM' to avoid NA collisions. Thanks to ntncmch.

no_counties_grid1: Grid of counties of Norway. Thanks to NanAmalie1.

tr_provinces_grid1: Grid of Provinces of Turkey. Thanks to sadettindemirel.

us_eastern_states_grid1: US State_grid showing only the states for which some portion falls east of the 100th meridian. Thanks to mihiarc.

br_states_grid3: Grid map of States of Brazil organized horizontally by region. Thanks to lzrmr.

us_states_territories_grid2: Grid map of US States and Territories. Thanks to edavidaja.

us_state_grid8: Grid map of US States including Washington, D.C.. Thanks to krothkin.

us_state_grid9: Grid map of US States including Washington, D.C.. Thanks to krothkin.

fr_departements_grid3: Grid of Departements of France. Thanks to leachareyre.

in_state_ut_grid3: Grid of Indian States and Union Territories. Thanks to aeschuma.

th_provinces_grid1: Grid of Provinces of Thailand. Thanks to PaulApivat.

th_bangkok_districts_grid1: Grid of Districts of Bangkok, Thailand. Thanks to PaulApivat.

ca_us_prov_state_grid1: Grid of Provinces and States of Canada and US. Thanks to ahcyip.

sy_governorates_grid1: Grid of ADM-1 units (governorates) for Syria. Thanks to jschon637.

ro_counties_grid1: Grid of counties (județe) of Romania. Thanks to bogdanantonescu.

us_va_health_districts_grid1: Grid of Commonwealth of Virginia's Health Districts. Thanks to osesusan.

us_state_without_DC_canada_prov_grid1: US States without DC with 10 Canadian Provinces. Thanks to opus1993.

ir_provinces_grid1: Provinces of Iran. Thanks to mcnakhaee.

co_departments_grid1: Departments of Colombia. Thanks to mikafui2020.

ir_tehran_districts_grid1: Grid for 22 districts in Tehran, Iran. Thanks to mcnakhaee.

ro_counties_grid2: Grid of Romanian Counties. Thanks to alexfg.

gb_sct_council_areas_grid1: Council Areas of Scotland. Thanks to AndrewAiton.

mw_districts_grid1: Grid of 27 districts of Malawi (all districts other than Likoma, which is an island). Thanks to taylorokonek.

dk_cph_grid1: Grid of the 10 districts of Copenhagen, Denmark. Thanks to Straubinger.

us_nv_counties_grid1: Counties of State of Nevada, United States. Thanks to schmidtDETR.

ie_counties_grid2: Counties of Ireland. Thanks to cbhurley.

bo_departments_grid1: Grid of Departamental Political Division of Bolivia. Thanks to tartagalensis.

ie_counties_grid3: Grid of 32 Counties of Ireland with Tipperary as one county. Thanks to superboreen.

co_departments_grid2: Grid of Departments of Colombia with municipal DANE codes. Thanks to sfanchi.

americas_countries_grid1: Map of countries of North and South America. Thanks to Aminsn.

us_census_div_grid2: Geofacet grid for the cartographic boundary map of the four official US regions, according to the US Census Bureau. Thanks to AnneIoannides.

ar_buenosaires_conurbano_grid1: Grids representing the greater Buenos Aires area which includes the autonomous city of Buenos Aires and some department/partidos of the Buenos Aires Province. Thanks to SantiagoEsteban.

us_md_counties_grid1: Counties of the State of Maryland, United States. Thanks to tecason.

us_ia_counties_grid2: Counties of the State of Iowa, United States. Thanks to tecason.

us_western_states_grid1: Grid of Western United States. Thanks to JHasselbeck.

us_ga_counties_grid1: Grid of Counties of Georgia, United States. Thanks to pjm407.

ro_counties_grid3: Counties (Judete) of Romania. Thanks to vparvu.

es_autonomous_communities_grid2: Autonomous Communities of Spain. Thanks to dieghernan.

qa_municipalities_grid1: Municipalities of Qatar. Thanks to stedy.

kz_region_grid1: Regions of Kazakhstan. Thanks to Sofiya2809.

middle_east_grid1: Countries of the Middle East with ISO codes. Thanks to anamartincalde.

cz_prague_districts_grid1: Grid of the 57 city districts of the city of Prague, Czech Republic using RUIAN IDs. Thanks to petrbouchal.

es_catalonia_comarques_grid1: Grid of Comarques of Catalonia, Soain. Thanks to Storydata.

no_counties_grid2: Counties of Norway in 2020. Thanks to LIVF.

fi_helsinki_neighborhoods_grid1: Neighborhoods of Helsinki, Finland. Thanks to msutela.

us_state_without_DC_AK_HI_grid1: Updated US state Without DC, AK and HI. Thanks to KWB4484.

ru_federal_subjects_grid2: Federal Subjects of Russia. Thanks to katerinaalex.

us_ut_counties_grid1: Grid of Counties of Utah, United States. Thanks to raogeorge.

in_state_ut_grid4: India states and union territories facet grid. Changes compared to previous grid: now the following union territories are included as well: Chandigarh, Delhi, Lakshadweep, Andaman and Nicobar Islands, Ladakh.. Thanks to rishabhshah-92.

ec_prov_grid2: Denser grid for Provinces of Ecuador including Galápagos Islands. Thanks to temporalista.

es_prov_grid1: Grid of Provinces of Spain. Thanks to EstebanAnd.

americas_countries_grid2: Grid of Countries of North and South America. Thanks to MathildeMousset.

us_nm_counties_grid1: Grid of Counties of New Mexico, United States. Thanks to ty-sanders.

us_me_counties_grid1: Grid of Counties of Maine, United States. Thanks to dikbrown.

cu_prov_grid1: Grid of Provinces of Cuba and Isla de la Juventud. Thanks to accostales.

ge_regions_grid1: Grid of Regions of Georgia. Thanks to nasrashvilin.

co_departments_grid3: Grid of Departments of Colombia. Thanks to JOTOR.

gb_london_boroughs_grid2: Boroughs of London with name wrapped to fit grid. Thanks to ICAITC.

gb_sct_aberdeenshire_IZ_grid1: Grid of Scottish intermediate zones for Aberdeenshire. Thanks to waynegault.

gb_sct_aberdeenshire_wards_grid1: Grid of Aberdeenshire, Scotland's Multi-member Wards. Thanks to waynegault.

at_vienna_districts_grid1: Districts of Vienna, Austria. Thanks to romanseidl.

gh_regions_grid1: Regions of Ghana. Thanks to Abu-sakara.

uy_departamentos_grid2: Departamentos of Uruguay. Thanks to RichDeto.

ch_vaud_districts_grid1: Districts for Canton of Vaud, Switzerland. Thanks to vdes2020.

us_ca_counties_FIPS_grid1: Grid of Counties of California, United States with FIPS codes. Thanks to ahcyip.

kr_seoul_district_grid2: Grid of Districts of Seoul, South Korea. Thanks to kjhnav.

kr_districts_grid1: Administrative Districts of South Korea. Thanks to kjhnav.

oecd_grid2: Grid of member countries of Organisation for Economic Co-operation and Development. Thanks to kjhnav.

bo_departments_grid2: Grid of Departamentos of Bolivia. Thanks to ccsuehara.

ca_us_prov_state_grid2: Grid of Provinces of Canada and States of United States. Thanks to semerson77.

us_il_counties_grid1: Grid of Counties of Illinois, United States. Thanks to amsutton.

tn_governorates_grid1: Grid of Governorates of Tunisia. Thanks to AminGhrabi.

gb_sct_glasgow_wards_grid1: Electoral Wards of Glasgow, Scotland. Thanks to nrennie.

kr_provinces_grid1: Provinces of Republic of Korea. Thanks to statkclee.

kr_counties_districts_cities_grid1: South Korean metropolitan cities' districts (gu), municipal cities (si), and counties (gun). Thanks to chichead.

us_ok_counties_grid1: Grid of Counties of Oklahoma, United States. Thanks to andrewvanleuven.

us_dc_neighborhoods_grid1: Grid of Neighborhoods of District of Columbia, United States. Thanks to rexarski.

us_mn_counties_grid1: Grid of Counties of Minnesota. Thanks to chadrent.

us_state_grid1

us_state_grid2

eu_grid1

aus_grid1

sa_prov_grid1

gb_london_boroughs_grid

nhs_scot_grid

india_grid1

india_grid2

argentina_grid1

br_states_grid1

sea_grid1

mys_grid1

fr_regions_grid1

de_states_grid1

us_or_counties_grid1

us_wa_counties_grid1

us_in_counties_grid1

us_in_central_counties_grid1

se_counties_grid1

sf_bay_area_counties_grid1

ua_region_grid1

mx_state_grid1

mx_state_grid2

scotland_local_authority_grid1

us_state_without_DC_grid1

italy_grid1

italy_grid2

be_province_grid1

us_state_grid3

jp_prefs_grid1

ng_state_grid1

bd_upazila_grid1

spain_prov_grid1

ch_cantons_grid1

ch_cantons_grid2

china_prov_grid1

world_86countries_grid

se_counties_grid2

uk_regions1

us_state_contiguous_grid1

sk_province_grid1

ch_aargau_districts_grid1

jo_gov_grid1

es_autonomous_communities_grid1

spain_prov_grid2

world_countries_grid1

br_states_grid2

china_city_grid1

kr_seoul_district_grid1

nz_regions_grid1

sl_regions_grid1

us_census_div_grid1

ar_tucuman_province_grid1

us_nh_counties_grid1

china_prov_grid2

pl_voivodeships_grid1

us_ia_counties_grid1

us_id_counties_grid1

ar_cordoba_dep_grid1

us_fl_counties_grid1

ar_buenosaires_communes_grid1

nz_regions_grid2

oecd_grid1

ec_prov_grid1

nl_prov_grid1

ca_prov_grid1

us_nc_counties_grid1

mx_ciudad_prov_grid1

bg_prov_grid1

us_hhs_regions_grid1

tw_counties_grid1

tw_counties_grid2

af_prov_grid1

us_mi_counties_grid1

pe_prov_grid1

sa_prov_grid2

mx_state_grid3

cn_bj_districts_grid1

us_va_counties_grid1

us_mo_counties_grid1

cl_santiago_prov_grid1

us_tx_capcog_counties_grid1

sg_planning_area_grid1

in_state_ut_grid1

cn_fujian_prov_grid1

ca_quebec_electoral_districts_grid1

nl_prov_grid2

cn_bj_districts_grid2

ar_santiago_del_estero_prov_grid1

ar_formosa_prov_grid1

ar_chaco_prov_grid1

ar_catamarca_prov_grid1

ar_jujuy_prov_grid1

ar_neuquen_prov_grid1

ar_san_luis_prov_grid1

ar_san_juan_prov_grid1

ar_santa_fe_prov_grid1

ar_la_rioja_prov_grid1

ar_mendoza_prov_grid1

ar_salta_prov_grid1

ar_rio_negro_prov_grid1

uy_departamentos_grid1

ar_buenos_aires_prov_electoral_dist_grid1

europe_countries_grid1

argentina_grid2

us_state_without_DC_grid2

jp_prefs_grid2

na_regions_grid1

mm_state_grid1

us_state_with_DC_PR_grid1

fr_departements_grid1

ar_salta_prov_grid2

ie_counties_grid1

sg_regions_grid1

us_ny_counties_grid1

ru_federal_subjects_grid1

us_ca_counties_grid1

lk_districts_grid1

us_state_without_DC_grid3

co_cali_subdivisions_grid1

us_in_northern_counties_grid1

italy_grid3

us_state_with_DC_PR_grid2

us_state_grid7

sg_planning_area_grid2

ch_cantons_fl_grid1

europe_countries_grid2

us_states_territories_grid1

us_tn_counties_grid1

us_il_chicago_community_areas_grid1

us_state_with_DC_PR_grid3

in_state_ut_grid2

at_states_grid1

us_pa_counties_grid1

us_oh_counties_grid1

fr_departements_grid2

us_wi_counties_grid1

africa_countries_grid1

no_counties_grid1

tr_provinces_grid1

us_eastern_states_grid1

br_states_grid3

us_states_territories_grid2

us_state_grid8

us_state_grid9

fr_departements_grid3

in_state_ut_grid3

th_provinces_grid1

th_bangkok_districts_grid1

ca_us_prov_state_grid1

sy_governorates_grid1

ro_counties_grid1

us_va_health_districts_grid1

us_state_without_DC_canada_prov_grid1

ir_provinces_grid1

co_departments_grid1

ir_tehran_districts_grid1

ro_counties_grid2

gb_sct_council_areas_grid1

mw_districts_grid1

dk_cph_grid1

us_nv_counties_grid1

ie_counties_grid2

bo_departments_grid1

ie_counties_grid3

co_departments_grid2

americas_countries_grid1

us_census_div_grid2

ar_buenosaires_conurbano_grid1

us_md_counties_grid1

us_ia_counties_grid2

us_western_states_grid1

us_ga_counties_grid1

ro_counties_grid3

es_autonomous_communities_grid2

qa_municipalities_grid1

kz_region_grid1

middle_east_grid1

cz_prague_districts_grid1

es_catalonia_comarques_grid1

no_counties_grid2

fi_helsinki_neighborhoods_grid1

us_state_without_DC_AK_HI_grid1

ru_federal_subjects_grid2

us_ut_counties_grid1

in_state_ut_grid4

ec_prov_grid2

es_prov_grid1

americas_countries_grid2

us_nm_counties_grid1

us_me_counties_grid1

cu_prov_grid1

ge_regions_grid1

co_departments_grid3

gb_london_boroughs_grid2

gb_sct_aberdeenshire_IZ_grid1

gb_sct_aberdeenshire_wards_grid1

at_vienna_districts_grid1

gh_regions_grid1

uy_departamentos_grid2

ch_vaud_districts_grid1

us_ca_counties_FIPS_grid1

kr_seoul_district_grid2

kr_districts_grid1

oecd_grid2

bo_departments_grid2

ca_us_prov_state_grid2

us_il_counties_grid1

tn_governorates_grid1

gb_sct_glasgow_wards_grid1

kr_provinces_grid1

kr_counties_districts_cities_grid1

us_ok_counties_grid1

us_dc_neighborhoods_grid1

us_mn_counties_grid1