Geo Grids

There are now 108 grids available in this package and more online. To view a full list of available grids, see here. To create and submit your own grid, see here. To see several examples of grids being used to visualize data, see facet_geo.

us_state_grid1: Grid layout for US states (including DC) Image reference here.

us_state_grid2: Grid layout for US states (including DC) Image reference here.

eu_grid1: Grid layout for the 28 EU Countries Image reference here.

aus_grid1: Grid layout for the Australian States and Territories. Image reference here. Thanks to jonocarroll.

sa_prov_grid1: Grid layout for the provinces of South Africa Image reference here. Thanks to jonmcalder.

gb_london_boroughs_grid: Grid layout for the boroughs of London. Note that the column code_ons contains the codes used by UK Office for National Statistics. Image reference here. Thanks to eldenvo.

nhs_scot_grid: Grid layout for a grid of NHS Scotland Health Boards. Note that the column code contains the codes used by UK Office for National Statistics. Image reference here. Thanks to jsphdms.

india_grid1: Grid layout for India states (not including union territories). Image reference here. Thanks to meysubb.

india_grid2: Grid layout for India states (not including union territories). Image reference here.

argentina_grid1: Grid for the 23 provinces of Argentina. It includes the Malvinas/Falkland Islands and the Antarctic Territories (these are disputed, but they are included since many researchers might use data from these locations). Image reference here. Thanks to eliocamp.

br_states_grid1: Grid for the 27 states of Brazil. Image reference here. Thanks to italocegatta.

sea_grid1: Grid for South East Asian countries. Image reference here. Thanks to jasonjb82.

mys_grid1: Grid for Malaysian states and territories. Image reference here. Thanks to jasonjb82.

fr_regions_grid1: Land and overseas regions of France. Codes are INSEE codes. Image reference here. Thanks to mtmx.

de_states_grid1: Grid for the German states ('Länder') Image reference here. Thanks to DominikVogel.

us_or_counties_grid1: Grid for Oregon counties. Image reference here. Thanks to aosmith16.

us_wa_counties_grid1: Grid for Washington counties. Image reference here.

us_in_counties_grid1: Grid for Indiana counties. Image reference here. Thanks to nateapathy.

us_in_central_counties_grid1: Grid for central Indiana counties. Image reference here. Thanks to nateapathy.

se_counties_grid1: Grid for counties of Sweden. Image reference here. Thanks to duleise.

sf_bay_area_counties_grid1: Grid of the 9 San Francisco Bay Area counties. Image reference here. Thanks to Eunoia.

ua_region_grid1: Grid of administrative divisions of Ukraine (24 oblasts, one autonomous region, and two cities). Image reference here. Thanks to woldemarg.

mx_state_grid1: Grid layout for the states of Mexico. Image reference here. Thanks to ikashnitsky.

mx_state_grid2: Grid layout for the states of Mexico. Image reference here. Thanks to diegovalle.

scotland_local_authority_grid1: Grid layout for the local authorities of Scotland. Image reference here. Thanks to davidhen.

us_state_grid3: Grid layout for US states (excluding DC) Image reference here. Thanks to ejr248.

italy_grid1: Grid layout for regions of Italy (in collaboration with Stella Cangelosi and Luciana Dalla Valle). Image reference here. Thanks to JulianStander.

italy_grid2: Grid layout for regions of Italy (in collaboration with Stella Cangelosi and Luciana Dalla Valle). Image reference here. Thanks to JulianStander.

be_province_grid1: Grid layout for provinces of Belgium plus Brussels, including names in three languages (French, Dutch, English) and Belgium internal codes (NIS). Image reference here. Thanks to ericlecoutre.

us_state_grid4: Grid layout for US states (including DC). Image reference here. Thanks to kanishkamisra.

jp_prefs_grid1: Grid layout for the prefectures of Japan. Image reference here. Thanks to uribo.

ng_state_grid1: Grid layout for the 37 Federal States of Nigeria. Image reference here. Thanks to ghosthedirewolf.

bd_upazila_grid1: Grid layout for Bangladesh 64 Upazilas. Image reference here. Thanks to ghosthedirewolf.

spain_prov_grid1: Grid layout for Provinces of Spain. Image reference here. Thanks to kintero.

ch_cantons_grid1: Grid layout for Cantons of Switzerland. Image reference here. Thanks to tinu-schneider.

ch_cantons_grid2: Grid layout for Cantons of Switzerland. Image reference here. Thanks to rastrau.

china_prov_grid1: Grid layout for Provinces of China. Image reference here. Thanks to weiyunna.

world_86countries_grid: Grid layout for 86 countries in the world. Image reference here. Thanks to akangsha.

se_counties_grid2: Grid for counties of Sweden. Image reference here. Thanks to richardohrvall.

uk_regions1: Grid for regions of the UK (aka EU standard NUTS 1 areas). Image reference here. Thanks to paulb20.

us_state_contiguous_grid1: Grid layout for the contiguous US states (including DC). Image reference here. Thanks to andrewsr.

sk_province_grid1: Grid layout for South Korean sis and dos (metropolitan/special/autonomous cities and provinces). Image reference here. Thanks to heon131.

ch_aargau_districts_grid1: Grid layout for Districts of the Canton of Aargau, Switzerland. Image reference here. Thanks to zumbov2.

jo_gov_grid1: Grid layout for Governorates of Jordan. Image reference here. Thanks to ghosthedirewolf.

spain_ccaa_grid1: Grid layout for Spanish 'Comunidades Autónomas'. Image reference here. Thanks to JoseAntonioOrtega.

spain_prov_grid2: Grid layout for Provinces of Spain. Image reference here. Thanks to JoseAntonioOrtega.

world_countries_grid1: Grid layout for countries of the world, with a few exclusions. See . Image reference here. Thanks to JoseAntonioOrtega.

br_states_grid2: Grid for the 27 states of Brazil. Image reference here. Thanks to hafen.

china_city_grid1: Grid layout of cities in China. Image reference here. Thanks to CharleneDeng1.

kr_seoul_district_grid1: Grid layout of Seoul's 25 districts. Image reference here. Thanks to yonghah.

nz_regions_grid1: Grid layout for regions of New Zealand. Image reference here. Thanks to pierreroudier.

sl_regions_grid1: Grid layout of Slovenian regions. Image reference here. Thanks to SR1986.

us_census_div_grid1: Grid layout of US Census divisions. Image reference here. Thanks to mkiang.

ar_tucuman_province_grid1: Grid layout for Argentina Tucumán Province political divisions (departments) Image reference here. Thanks to TuQmano.

us_nh_counties_grid1: Grid layout for the 10 counties in New Hampshire. Image reference here. Thanks to soungl.

china_prov_grid2: Grid layout for Provinces of China. Image reference here. Thanks to jw2531.

pl_voivodeships_grid1: Grid layout for Polish voivodeships (provinces) Image reference here. Thanks to erzk.

us_ia_counties_grid1: Grid layout for counties in Iowa Image reference here. Thanks to jrennyb.

us_id_counties_grid1: Grid layout for counties in Idaho Image reference here. Thanks to hathawayj.

ar_cordoba_dep_grid1: Grid layout for departments of Cordoba province in Argentina. Image reference here. Thanks to TuQmano.

us_fl_counties_grid1: Grid for Florida counties. Image reference here. Thanks to ejr248.

ar_buenosaires_communes_grid1: Grid for communes of Buenos Aires, Argentina. Image reference here. Thanks to TuQmano.

nz_regions_grid2: Grid layout for regions of New Zealand. Image reference here. Thanks to pierreroudier.

oecd_grid1: Grid layout for OECD member countries. Image reference here. Thanks to arcruz0.

ec_prov_grid1: Grid layout for provinces of Ecuador Image reference here. Thanks to Ricardo95RM.

nl_prov_grid1: Grid layout for provinces of Netherlands Image reference here. Thanks to ruditurksema.

ca_prov_grid1: Grid layout for provinces of Canada Image reference here. Thanks to michael-chong.

us_nc_counties_grid1: Grid layout for Counties of North Carolina, United States Image reference here. Thanks to mtdukes.

mx_ciudad_prov_grid1: Grid layout for Districts of Mexico City, Mexico Image reference here. Thanks to Ivangea.

bg_prov_grid1: Grid layout for provinces of Bulgaria Image reference here. Thanks to savinastoitsova.

us_hhs_regions_grid1: This grid approximates the U.S. Health and Human Services Region map Image reference here. Thanks to akitepowell.

tw_counties_grid1: Grid layout for Counties of Taiwan Image reference here. Thanks to csh484912274.

tw_counties_grid2: Grid layout for Counties of Taiwan including Lienchiang County Image reference here. Thanks to csh484912274.

af_prov_grid1: Grid layout for Provinces of Afghanistan Image reference here. Thanks to jrennyb.

us_mi_counties_grid1: Grid layout for Counties of Michigan, United States Image reference here. Thanks to jrennyb.

pe_prov_grid1: Grid layout for Provinces of Peru Image reference here. Thanks to jmcastagnetto.

sa_prov_grid2: Grid layout for Provinces of South Africa Image reference here. Thanks to kamermanpr.

mx_state_grid3: Grid layout for States of Mexico Image reference here. Thanks to ikashnitsky.

cn_bj_districts_grid1: Grids for Administrative Districts of Beijing, China Image reference here. Thanks to shiedelweiss.

us_va_counties_grid1: Grids for Counties of Virginia, United States Image reference here. Thanks to joshyazman.

us_mo_counties_grid1: Grids for Counties of Missouri, United States Image reference here. Thanks to Yanqi-Xu.

cl_santiago_prov_grid1: Communes of Santiago Province, Chile Image reference here. Thanks to robsalasco.

us_tx_capcog_counties_grid1: This is a grid of a 10 county planning region around Austin, Texas, United States Image reference here. Thanks to mth444.

sg_planning_area_grid1: Grids for Planning Areas of Singapore Image reference here. Thanks to Elenafuyi.

in_state_ut_grid1: Grid of Indian States and Union Territories Image reference here. Thanks to seanangio.

cn_fujian_prov_grid1: Grid of counties of Fujian Province, China Image reference here. Thanks to nannanchen333.

ca_quebec_electoral_districts_grid1: Grid of Electoral Districts of Québec, Canada Image reference here. Thanks to jhroy.

nl_prov_grid2: Grid with the provinces of The Netherlands with codes that are used by the statistical institute of NL Image reference here. Thanks to edwindj.

cn_bj_districts_grid2: Grid with districts of Beijing, China Image reference here. Thanks to zouhx11.

ar_santiago_del_estero_prov_grid1: Grid with districts of Santiago del Estero Province, Argentina Image reference here. Thanks to TuQmano.

ar_formosa_prov_grid1: Grid with districts of Formosa Province, Argentina Image reference here. Thanks to TuQmano.

ar_chaco_prov_grid1: Grid with districts of Chaco Province, Argentina Image reference here. Thanks to TuQmano.

ar_catamarca_prov_grid1: Grid with districts of Catamarca Province, Argentina Image reference here. Thanks to TuQmano.

ar_jujuy_prov_grid1: Grid with districts of Jujuy Province, Argentina Image reference here. Thanks to TuQmano.

ar_neuquen_prov_grid1: Grid with districts of Neuquen Province, Argentina Image reference here. Thanks to TuQmano.

ar_san_luis_prov_grid1: Grid with districts of San Luis Province, Argentina Image reference here. Thanks to TuQmano.

ar_san_juan_prov_grid1: Grid with districts of San Juan Province, Argentina Image reference here. Thanks to TuQmano.

ar_santa_fe_prov_grid1: Grid with districts of Santa Fe Province, Argentina Image reference here. Thanks to TuQmano.

ar_la_rioja_prov_grid1: Grid with districts of La Rioja Province, Argentina Image reference here. Thanks to TuQmano.

ar_mendoza_prov_grid1: Grid with districts of Mendoza Province, Argentina Image reference here. Thanks to TuQmano.

ar_salta_prov_grid1: Grid with districts of Salta Province, Argentina Image reference here. Thanks to TuQmano.

ar_rio_negro_prov_grid1: Grid with districts of Rio Negro Province, Argentina Image reference here. Thanks to TuQmano.

uy_departamentos_grid1: Grid with Departamentos of Uruguay Image reference here. Thanks to TuQmano.

ar_buenos_aires_prov_electoral_dist_grid1: Grid with Electoral Districts of Buenos Aires Province, Argentina Image reference here. Thanks to TuQmano.

europe_countries_grid1: Grid layout for all European countries except Vatican City, Monaco, San Marino and Liechtenstein Image reference here. Thanks to THargreaves.

argentina_grid2: Grid layout for Argentina without Islas Malvinas and Antártida Argentina Image reference here. Thanks to TuQmano.

us_state_grid5: Grid layout for United States with AK and HI flush with CA Image reference here. Thanks to christophercannon.

jp_prefs_grid2: Grid layout for Prefectures of Japan Image reference here. Thanks to Ryo-N7.

us_state_grid1

us_state_grid2

eu_grid1

aus_grid1

sa_prov_grid1

gb_london_boroughs_grid

nhs_scot_grid

india_grid1

india_grid2

argentina_grid1

br_states_grid1

sea_grid1

mys_grid1

fr_regions_grid1

de_states_grid1

us_or_counties_grid1

us_wa_counties_grid1

us_in_counties_grid1

us_in_central_counties_grid1

se_counties_grid1

sf_bay_area_counties_grid1

ua_region_grid1

mx_state_grid1

mx_state_grid2

scotland_local_authority_grid1

us_state_grid3

italy_grid1

italy_grid2

be_province_grid1

us_state_grid4

jp_prefs_grid1

ng_state_grid1

bd_upazila_grid1

spain_prov_grid1

ch_cantons_grid1

ch_cantons_grid2

china_prov_grid1

world_86countries_grid

se_counties_grid2

uk_regions1

us_state_contiguous_grid1

sk_province_grid1

ch_aargau_districts_grid1

jo_gov_grid1

spain_ccaa_grid1

spain_prov_grid2

world_countries_grid1

br_states_grid2

china_city_grid1

kr_seoul_district_grid1

nz_regions_grid1

sl_regions_grid1

us_census_div_grid1

ar_tucuman_province_grid1

us_nh_counties_grid1

china_prov_grid2

pl_voivodeships_grid1

us_ia_counties_grid1

us_id_counties_grid1

ar_cordoba_dep_grid1

us_fl_counties_grid1

ar_buenosaires_communes_grid1

nz_regions_grid2

oecd_grid1

ec_prov_grid1

nl_prov_grid1

ca_prov_grid1

us_nc_counties_grid1

mx_ciudad_prov_grid1

bg_prov_grid1

us_hhs_regions_grid1

tw_counties_grid1

tw_counties_grid2

af_prov_grid1

us_mi_counties_grid1

pe_prov_grid1

sa_prov_grid2

mx_state_grid3

cn_bj_districts_grid1

us_va_counties_grid1

us_mo_counties_grid1

cl_santiago_prov_grid1

us_tx_capcog_counties_grid1

sg_planning_area_grid1

in_state_ut_grid1

cn_fujian_prov_grid1

ca_quebec_electoral_districts_grid1

nl_prov_grid2

cn_bj_districts_grid2

ar_santiago_del_estero_prov_grid1

ar_formosa_prov_grid1

ar_chaco_prov_grid1

ar_catamarca_prov_grid1

ar_jujuy_prov_grid1

ar_neuquen_prov_grid1

ar_san_luis_prov_grid1

ar_san_juan_prov_grid1

ar_santa_fe_prov_grid1

ar_la_rioja_prov_grid1

ar_mendoza_prov_grid1

ar_salta_prov_grid1

ar_rio_negro_prov_grid1

uy_departamentos_grid1

ar_buenos_aires_prov_electoral_dist_grid1

europe_countries_grid1

argentina_grid2

us_state_grid5

jp_prefs_grid2